test: move VaultClientError test to common tests + cleanup relations
